Finalize mypage
