+++ frontend/src/javascripts/components/index/user-only.js
@@ -1,12 +1,10 @@
 import React, { Component } from 'react'
 import { Route, Switch, Redirect } from 'react-router-dom'
 import { connect } from 'react-redux'
-import Header from '../organisms/header'
-import Mypage from '../organisms/mypage-main'
+import Mypage from '../pages/users/mypage'
 import ProjectPage from '../organisms/project-page-main'
 import HistoryPage from '../organisms/history-page-main'
 import ProfilePageMain from '../organisms/profile-page-main'
-import Footer from '../organisms/footer'
 
 class UserOnly extends Component {
   render() {
@@ -17,7 +15,6 @@ class UserOnly extends Component {
 
     return currentUser ? (
       <div className="page-container">
-        <Header {...this.props} />
         <Switch>
           <Route
             exact
@@ -37,7 +34,6 @@ class UserOnly extends Component {
           <Route exact path={`${url}/:userId`} component={Mypage} />
           <Route render={() => <h2>404 Not Found</h2>} />
         </Switch>
-        <Footer {...this.props} />
       </div>
     ) : (
       <Redirect to="/guests" />
